relcheck_v3/eval/requirements.txt
pydantic>=2.0
pycocoevalcap
transformers
bitsandbytes
accelerate
tqdm
pandas
Pillow
hypothesis
pytest
